OAuth2 setup moved to middleware package
