+++ api/cmd/serve.go
@@ -14,10 +14,6 @@ import (
 func createHTTPServerOptions(c server.Config) ([]server.HTTPServerOption, error) {
 	opts := make([]server.HTTPServerOption, 0)
 
-	opts = append(opts,
-		server.WithOAuth2(c.OAuth2Option),
-	)
-
 	if c.Profiling {
 		opts = append(
 			opts,
