+++ api/server/http.go
@@ -5,17 +5,14 @@ import (
 	"net/http"
 	pprof "net/http/pprof"
 
-	jwt "github.com/dgrijalva/jwt-go"
 	_ "github.com/equinor/oneseismic/api/docs" // docs is generated by Swag CLI, you have to import it.
 	l "github.com/equinor/oneseismic/api/logger"
 	"github.com/equinor/oneseismic/api/middleware"
 	claimsmiddleware "github.com/equinor/oneseismic/api/middleware/claims"
-	jwtmiddleware "github.com/iris-contrib/middleware/jwt"
 	prometheusmiddleware "github.com/iris-contrib/middleware/prometheus"
 	"github.com/iris-contrib/swagger/v12"
 	"github.com/iris-contrib/swagger/v12/swaggerFiles"
 	"github.com/kataras/iris/v12"
-	irisCtx "github.com/kataras/iris/v12/context"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 )
 
@@ -45,6 +42,12 @@ func Create(c Config) (*HTTPServer, error) {
 		app:                app,
 		hostAddr:           c.HostAddr}
 
+	auth, _ := middleware.Oauth2(c.OAuth2Option)
+	hs.app.Use(auth)
+
+	claimsHandler := claimsmiddleware.New(c.OAuth2Option.Audience, c.OAuth2Option.Issuer)
+	hs.app.Use(claimsHandler.Validate)
+
 	return &hs, nil
 }
 
@@ -60,67 +63,6 @@ func Configure(hs *HTTPServer, opts ...HTTPServerOption) error {
 	hs.registerEndpoints()
 
 	return nil
-}
-
-func WithOAuth2(oauthOpt OAuth2Option) HTTPServerOption {
-
-	return newFuncOption(func(hs *HTTPServer) error {
-		sigKeySet, err := middleware.GetOIDCKeySet(oauthOpt.AuthServer)
-		if err != nil {
-			return fmt.Errorf("Couldn't get keyset: %v", err)
-		}
-
-		rsaJWTHandler := jwtmiddleware.New(jwtmiddleware.Config{
-			ValidationKeyGetter: func(t *jwt.Token) (interface{}, error) {
-
-				if t.Method.Alg() != "RS256" {
-					return nil, fmt.Errorf("unexpected jwt signing method=%v", t.Header["alg"])
-				}
-				return sigKeySet[t.Header["kid"].(string)], nil
-
-			},
-			ContextKey:    "user-jwt",
-			SigningMethod: jwt.SigningMethodRS256,
-		})
-
-		onRS256Pass := func(ctx irisCtx.Context, err error) {
-
-			if err == nil || err.Error() == "unexpected jwt signing method=RS256" {
-				return
-			}
-			jwtmiddleware.OnError(ctx, err)
-		}
-		hmacJWTHandler := jwtmiddleware.New(jwtmiddleware.Config{
-			ValidationKeyGetter: func(t *jwt.Token) (interface{}, error) {
-
-				if t.Method.Alg() != "HS256" {
-					return nil, fmt.Errorf("unexpected jwt signing method=%v", t.Header["alg"])
-				}
-				return oauthOpt.APISecret, nil
-			},
-			ContextKey:    "service-jwt",
-			SigningMethod: jwt.SigningMethodHS256,
-			ErrorHandler:  onRS256Pass,
-		})
-
-		if len(oauthOpt.Issuer) == 0 {
-			oauthOpt.Issuer = oauthOpt.AuthServer.String()
-		}
-
-		claimsHandler := claimsmiddleware.New(oauthOpt.Audience, oauthOpt.Issuer)
-
-		auth := func(ctx irisCtx.Context) {
-			hmacJWTHandler.Serve(ctx)
-			serviceToken := ctx.Values().Get("service-jwt")
-			if serviceToken == nil {
-				rsaJWTHandler.Serve(ctx)
-			}
-
-		}
-		hs.app.Use(auth)
-		hs.app.Use(claimsHandler.Validate)
-		return nil
-	})
 }
 
 func (hs *HTTPServer) registerEndpoints() {
