+++ api/server/option.go
@@ -1,7 +1,5 @@
 package server
 
-import "net/url"
-
 type EmptyOption struct{}
 
 func (EmptyOption) apply(*HTTPServer) (err error) { return }
@@ -19,11 +17,4 @@ func newFuncOption(f func(*HTTPServer) error) *funcOption {
 	return &funcOption{
 		f: f,
 	}
-}
-
-type OAuth2Option struct {
-	AuthServer *url.URL
-	Audience   string
-	Issuer     string
-	APISecret  []byte
 }
